view

+++ src/Controller/EnterpriseController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Enterprise;
+use App\Repository\EnterpriseRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,29 +13,58 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/enterprise')]
 class EnterpriseController extends AbstractController
 {
-    #[Route('/{id}', name: 'app_enterprise_view', methods: ['GET'])]
-    public function view(Enterprise $enterprise): Response
+    #[Route('/{num}', name: 'app_enterprise_view', methods: ['GET'])]
+    public function view(string $num, EnterpriseRepository $repo): Response
     {
-        
+        $enterprise = $repo->findByEnterpriseNumber($num);
+
+        if (!$enterprise) {
+            throw $this->createNotFoundException("Entreprise introuvable");
+        }
+
+        return $this->render('enterprise/view.html.twig', [
+            'enterprise' => $enterprise
+        ]);
     }
 
-    #[Route('/add', name: 'app_enterprise_add', methods: ['GET', 'POST'])]
-    public function add(Request $request, EntityManagerInterface $em): Response
+    #[Route('/{num}/edit', name: 'app_enterprise_edit', methods: ['GET', 'POST'])]
+    public function edit(string $num, Request $request, EnterpriseRepository $repo, EntityManagerInterface $em): Response
     {
-        
+        $enterprise = $repo->findByEnterpriseNumber($num);
+
+        if (!$enterprise) {
+            throw $this->createNotFoundException("Entreprise introuvable");
+        }
+
+        if ($request->isMethod('POST')) {
+            $enterprise->setStatus($request->request->get('Status'));
+            $enterprise->setJuridicalForm($request->request->get('JuridicalForm'));
+
+            if ($request->request->get('StartDate')) {
+                $enterprise->setStartDate(new \DateTime($request->request->get('StartDate')));
+            }
+
+            $em->flush();
+
+            return $this->redirectToRoute('app_enterprise_view', [
+                'num' => $enterprise->getEnterpriseNumber()
+            ]);
+        }
+
+        return $this->render('enterprise/edit.html.twig', [
+            'enterprise' => $enterprise
+        ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_enterprise_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Enterprise $enterprise, EntityManagerInterface $em): Response
+    #[Route('/{num}/delete', name: 'app_enterprise_delete', methods: ['GET'])]
+    public function delete(string $num, EnterpriseRepository $repo, EntityManagerInterface $em): Response
     {
-        
-    }
+        $enterprise = $repo->findByEnterpriseNumber($num);
 
-    #[Route('/{id}/delete', name: 'app_enterprise_delete', methods: ['GET'])]
-    public function delete(EntityManagerInterface $em, Enterprise $enterprise): Response
-    {
-        $em->remove($enterprise);
-        $em->flush();
+        if ($enterprise) {
+            $em->remove($enterprise);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
